Switches to dark theme
Redesigning the app with a dark theme. Blue gray is now used as the
primary color and a black background is used in the palette screen.
Swatches have been adjusted to take the whole bottom box in the grid
view, and to display as a grid in the palette view.